+++ src/Cassandra.Tests/ConnectionTests.cs
@@ -22,7 +22,7 @@ using System.Net;
 using System.Threading;
 using System.Threading.Tasks;
 using Cassandra.Connections;
-using Cassandra.Observers;
+using Cassandra.Observers.Null;
 using Cassandra.Requests;
 using Cassandra.Responses;
 using Cassandra.Serialization;
