+++ src/Cassandra.Tests/ExecutionProfiles/RequestHandlerTests.cs
@@ -24,7 +24,7 @@ using System.Threading.Tasks;
 
 using Cassandra.Connections;
 using Cassandra.ExecutionProfiles;
-using Cassandra.Observers;
+using Cassandra.Observers.Null;
 using Cassandra.Requests;
 using Cassandra.Responses;
 using Cassandra.Serialization;
