+++ src/Cassandra.Tests/HostConnectionPoolTests.cs
@@ -25,7 +25,7 @@ using Cassandra.Connections;
 using Cassandra.Metrics;
 using Cassandra.Metrics.Internal;
 using Cassandra.Metrics.Providers.Null;
-using Cassandra.Observers;
+using Cassandra.Observers.Null;
 using Cassandra.Requests;
 using Cassandra.Serialization;
 using Cassandra.Tasks;
